requirements.txt
annotated-types==0.7.0
anyio==4.8.0
blinker==1.9.0
certifi==2024.12.14
charset-normalizer==3.4.0
click==8.1.8
colorama==0.4.6
diskcache==5.6.3
Flask==3.1.0
Flask-Cors==5.0.0
gpt4all==2.8.2
gunicorn==23.0.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
iniconfig==2.0.0
itsdangerous==2.2.0
Jinja2==3.1.4
llama_cpp_python==0.3.5
MarkupSafe==3.0.2
MouseInfo==0.1.3
numpy==2.2.0
ollama==0.4.7
packaging==24.2
pika==1.3.2
pluggy==1.5.0
PyAutoGUI==0.9.54
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
PyMsgBox==1.0.9
pyperclip==1.9.0
PyRect==0.2.0
PyScreeze==1.0.1
pytest==8.3.4
python-dotenv==1.0.1
pytweening==1.2.0
requests==2.32.3
setuptools==75.6.0
sniffio==1.3.1
tqdm==4.67.1
typing_extensions==4.12.2
urllib3==2.2.3
Werkzeug==3.1.3
wheel==0.45.1
